+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Fixed
+
+- `Select`/`SelectMulti` keyboard navigation when filtering and going from > 100 to < 100 options
+
 ## [0.9.5] - 2020-07-01
 
 ### Added